requirements.txt
google-generativeai
pymongo
python-dotenv
tqdm
pyvis